requirements.txt
kafka-python
requests
langchain
langchain-community
langchain-qdrant
sentence-transformers
pypdf
